Move Reference Data script to common package
